requirements.txt
streamlit
openai-whisper
numpy
scipy 